pyproject.toml
[project]
name = "sql-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.135.1",
    "openpyxl>=3.1.5",
    "sqlalchemy>=2.0",
    "psycopg2-binary>=2.9",
    "openai>=2.26.0",
    "postgres-mcp>=0.3.0",
    "streamlit>=1.44.0",
    "pandas<3",
]

[project.scripts]
sql-agent-mcp = "app.service.postgres_mcp:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
markers = [
    "eval: LLM evaluation test (slow, needs running server + DB)",
]
